Editor: Move editor toggle commands to the editor package
