+++ packages/edit-post/src/components/layout/index.js
@@ -47,7 +47,7 @@ import MetaBoxes from '../meta-boxes';
 import WelcomeGuide from '../welcome-guide';
 import { store as editPostStore } from '../../store';
 import { unlock } from '../../lock-unlock';
-import useCommonCommands from '../../hooks/commands/use-common-commands';
+import useEditPostCommands from '../../commands/use-commands';
 
 const { getLayoutStyles } = unlock( blockEditorPrivateApis );
 const { useCommands } = unlock( coreCommandsPrivateApis );
@@ -131,7 +131,7 @@ function useEditorStyles() {
 
 function Layout( { initialPost } ) {
 	useCommands();
-	useCommonCommands();
+	useEditPostCommands();
 
 	const isMobileViewport = useViewportMatch( 'medium', '<' );
 	const isWideViewport = useViewportMatch( 'large' );
