+++ packages/editor/src/components/commands/index.js
@@ -2,10 +2,14 @@
  * WordPress dependencies
  */
 import { useSelect, useDispatch } from '@wordpress/data';
-import { __ } from '@wordpress/i18n';
+import { __, isRTL } from '@wordpress/i18n';
 import {
+	blockDefault,
 	code,
+	drawerLeft,
+	drawerRight,
 	edit,
+	formatListBullets,
 	listView,
 	external,
 	keyboard,
@@ -38,6 +42,7 @@ function useEditorCommandLoader() {
 		isViewable,
 		isCodeEditingEnabled,
 		isRichEditingEnabled,
+		isPublishSidebarEnabled,
 	} = useSelect( ( select ) => {
 		const { get } = select( preferencesStore );
 		const { isListViewOpened, getCurrentPostType, getEditorSettings } =
@@ -56,8 +61,11 @@ function useEditorCommandLoader() {
 			isViewable: getPostType( getCurrentPostType() )?.viewable ?? false,
 			isCodeEditingEnabled: getEditorSettings().codeEditingEnabled,
 			isRichEditingEnabled: getEditorSettings().richEditingEnabled,
+			isPublishSidebarEnabled:
+				select( editorStore ).isPublishSidebarEnabled(),
 		};
 	}, [] );
+	const { getActiveComplementaryArea } = useSelect( interfaceStore );
 	const { toggle } = useDispatch( preferencesStore );
 	const { createInfoNotice } = useDispatch( noticesStore );
 	const {
@@ -66,7 +74,8 @@ function useEditorCommandLoader() {
 		switchEditorMode,
 		toggleDistractionFree,
 	} = useDispatch( editorStore );
-	const { openModal } = useDispatch( interfaceStore );
+	const { openModal, enableComplementaryArea, disableComplementaryArea } =
+		useDispatch( interfaceStore );
 	const { getCurrentPostId } = useSelect( editorStore );
 	const allowSwitchEditorMode = isCodeEditingEnabled && isRichEditingEnabled;
 
@@ -205,6 +214,57 @@ function useEditorCommandLoader() {
 					: __( 'Breadcrumbs visible.' ),
 				{
 					id: 'core/editor/toggle-breadcrumbs/notice',
+					type: 'snackbar',
+				}
+			);
+		},
+	} );
+
+	commands.push( {
+		name: 'core/open-settings-sidebar',
+		label: __( 'Toggle settings sidebar' ),
+		icon: isRTL() ? drawerLeft : drawerRight,
+		callback: ( { close } ) => {
+			const activeSidebar = getActiveComplementaryArea( 'core' );
+			close();
+			if ( activeSidebar === 'edit-post/document' ) {
+				disableComplementaryArea( 'core' );
+			} else {
+				enableComplementaryArea( 'core', 'edit-post/document' );
+			}
+		},
+	} );
+
+	commands.push( {
+		name: 'core/open-block-inspector',
+		label: __( 'Toggle block inspector' ),
+		icon: blockDefault,
+		callback: ( { close } ) => {
+			const activeSidebar = getActiveComplementaryArea( 'core' );
+			close();
+			if ( activeSidebar === 'edit-post/block' ) {
+				disableComplementaryArea( 'core' );
+			} else {
+				enableComplementaryArea( 'core', 'edit-post/block' );
+			}
+		},
+	} );
+
+	commands.push( {
+		name: 'core/toggle-publish-sidebar',
+		label: isPublishSidebarEnabled
+			? __( 'Disable pre-publish checks' )
+			: __( 'Enable pre-publish checks' ),
+		icon: formatListBullets,
+		callback: ( { close } ) => {
+			close();
+			toggle( 'core', 'isPublishSidebarEnabled' );
+			createInfoNotice(
+				isPublishSidebarEnabled
+					? __( 'Pre-publish checks disabled.' )
+					: __( 'Pre-publish checks enabled.' ),
+				{
+					id: 'core/editor/publish-sidebar/notice',
 					type: 'snackbar',
 				}
 			);
